requirements.txt
asgiref==3.2.10
Django==3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.1
PyJWT==1.7.1
pytz==2022.7.1
sqlparse==0.4.3
